+++ skate-plugin/src/main/kotlin/com/slack/sgp/intellij/modeltranslator/GenerateTranslatorBodyAction.kt
@@ -55,10 +55,7 @@ class GenerateTranslatorBodyAction(private val bundle: TranslatorBundle) : Inten
   private fun sendUsageTrace(project: Project, startTimestamp: Instant) {
     val skateSpanBuilder =
       SkateSpanBuilder().apply {
-        addSpanTag(
-          "event",
-          ModelTranslatorEvent.MODEL_TRANSLATOR_GENERATED,
-        )
+        addSpanTag("event", ModelTranslatorEvent.MODEL_TRANSLATOR_GENERATED)
       }
     SkateTraceReporter(project)
       .createPluginUsageTraceAndSendTrace(
