+++ skate-plugin/src/main/kotlin/com/slack/sgp/intellij/tracing/SkateSpanBuilder.kt
@@ -29,13 +29,14 @@ class SkateSpanBuilder {
   fun addSpanTag(key: String, event: SkateTracingEvent) {
     keyValueList.apply { key tagTo event.name }
   }
+
   fun addSpanTags(tags: Map<String, Any>) {
     tags.forEach { (key, value) ->
       when (value) {
-        is String -> keyValueList.apply { key tagTo value}
-        is Double -> keyValueList.apply { key tagTo value}
-        is Long -> keyValueList.apply { key tagTo value}
-        is Boolean -> keyValueList.apply { key tagTo value}
+        is String -> keyValueList.apply { key tagTo value }
+        is Double -> keyValueList.apply { key tagTo value }
+        is Long -> keyValueList.apply { key tagTo value }
+        is Boolean -> keyValueList.apply { key tagTo value }
         else -> throw IllegalArgumentException("Unsupported value type for key: ${key}")
       }
     }
